# platform/contracts/openapi/public-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Unified Software Commercialization Platform Public API",
    "version": "1.0.0-draft.1",
    "description": "Contract-first public client and admin API. Product, Application, Tenant and User contexts are resolved from verified sessions; identifiers supplied by clients are never authorization facts."
  },
  "servers": [{ "url": "/", "description": "Current environment" }],
...
  "paths": {
...
    "/api/v1/admin/products": {
      "get": {
        "tags": ["Product"], "operationId": "listProducts", "summary": "List products visible to the administrator",
        "security": [{ "AdminAccessCookie": [] }, { "AdminBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }],
        "responses": { "200": { "$ref": "#/components/responses/ProductList" }, "default": { "$ref": "#/components/responses/Error" } }
      },
      "post": {
        "tags": ["Product"], "operationId": "createProduct", "summary": "Create a product and its official tenant",
        "security": [{ "AdminAccessCookie": [] }, { "AdminBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }, { "$ref": "#/components/parameters/IdempotencyKey" }, { "$ref": "#/components/parameters/AdminCsrfToken" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProductRequest" } } } },
        "responses": { "201": { "$ref": "#/components/responses/Product" }, "default": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "tags": ["Identity"], "operationId": "listUsers", "summary": "List redacted global users for a platform administrator",
        "description": "This global collection requires platform scope. Product and tenant administrators must use their scoped workspace collection.",
        "x-required-permission": "identity.user.read", "x-required-scope": "platform",
        "security": [{ "AdminAccessCookie": [] }, { "AdminBearer": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RequestId" }, { "$ref": "#/components/parameters/UserQuery" }, { "$ref": "#/components/parameters/AccountStatusFilter" }, { "$ref": "#/components/parameters/Cursor" }, { "$ref": "#/components/parameters/PageSize" }],
        "responses": { "200": { "$ref": "#/components/responses/UserPage" }, "default": { "$ref": "#/components/responses/Error" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ClientSessionBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque", "description": "Short-lived client session containing server-resolved ProductContext, ApplicationContext and TenantContext." },
      "UserBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque", "description": "User session bound to trusted Product/Application and server-resolved Tenant context." },
      "HostedSessionCookie": { "type": "apiKey", "in": "cookie", "name": "__Host-platform_hosted_session", "description": "Short-lived Hosted UI browser session. Server sets Secure, HttpOnly, SameSite=Lax, Path=/ and no Domain. Unsafe methods also require exact Hosted Origin and X-CSRF-Token." },
      "AdminAccessCookie": { "type": "apiKey", "in": "cookie", "name": "__Host-platform_admin_access", "description": "Default browser administrator access. Server sets Secure, HttpOnly, SameSite=Strict, Path=/ and no Domain. Unsafe methods also require exact Origin and X-CSRF-Token." },
      "AdminRefreshCookie": { "type": "apiKey", "in": "cookie", "name": "__Secure-platform_admin_refresh", "description": "Rotating browser refresh proof. Server sets Secure, HttpOnly, SameSite=Strict, Path=/api/v1/admin/auth and no Domain; accepted only by admin refresh/logout." },
      "AdminBearer": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque", "description": "Optional non-browser administrator transport explicitly issued to controlled CLI/automation; every target product/tenant is re-authorized by Access Control." },
      "DeveloperKey": { "type": "http", "scheme": "bearer", "bearerFormat": "developer-key", "description": "Restricted developer key. The key resolves Product/Tenant/scope server-side and cannot select arbitrary contexts." }
    },
    "parameters": {
      "RequestId": { "name": "X-Request-Id", "in": "header", "required": false, "description": "Optional caller correlation ID; not an authorization or idempotency fact.", "schema": { "type": "string", "minLength": 8, "maxLength": 128 } },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "description": "Opaque operation-scoped idempotency key. Reuse with a different payload returns conflict.", "schema": { "type": "string", "minLength": 16, "maxLength": 128 } },
      "AdminCsrfToken": { "name": "X-CSRF-Token", "in": "header", "required": false, "description": "Required with exact Origin for unsafe requests authenticated by an administrator access Cookie, except refresh which recovers after access expiry using exact Origin plus SameSite refresh Cookie. Bound to the server session and rotated with it. Omit for Admin Bearer.", "schema": { "type": "string", "minLength": 32, "maxLength": 256 } },
...
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 512 } },
      "PageSize": { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } },
      "UserQuery": { "name": "query", "in": "query", "description": "User ID or display-name prefix, or an exact normalized email/phone identifier. Identifier values are never used for fuzzy matching.", "schema": { "type": "string", "minLength": 1, "maxLength": 320 } },
      "AccountStatusFilter": { "name": "account_status", "in": "query", "schema": { "type": "string", "enum": ["active", "locked", "disabled"] } },
...
    },
    "headers": {
      "RequestId": { "description": "Server request/correlation ID.", "schema": { "type": "string" } }
    },
    "responses": {
      "Error": { "description": "Stable safe error", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } },
...
      "Product": { "$ref": "#/components/responses/JsonProduct" },
      "JsonProduct": { "description": "Product", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
...
      "ProductList": { "description": "Authorized products", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/ProductSummary" } } }, "additionalProperties": false } } } },
...
      "UserPage": { "description": "Scoped redacted user page", "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserPage" } } } },
...
    },
    "schemas": {
      "Identifier": { "type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$" },
      "StableCode": { "type": "string", "minLength": 3, "maxLength": 64, "pattern": "^[a-z][a-z0-9]*(?:[._-][a-z0-9]+)*$" },
...
      "Timestamp": { "type": "string", "format": "date-time" },
...
      "ErrorEnvelope": { "type": "object", "required": ["type", "title", "status", "code", "request_id", "retryable"], "properties": { "type": { "type": "string", "format": "uri-reference" }, "title": { "type": "string" }, "status": { "type": "integer", "minimum": 400, "maximum": 599 }, "code": { "type": "string", "pattern": "^[A-Za-z0-9_.-]+$" }, "detail": { "type": "string" }, "request_id": { "type": "string" }, "retryable": { "type": "boolean" }, "retry_after_seconds": { "type": "integer", "minimum": 0 }, "field_errors": { "type": "array", "items": { "type": "object", "required": ["field", "code"], "properties": { "field": { "type": "string" }, "code": { "type": "string" }, "message": { "type": "string" } } } } } },
...
      "CreateProductRequest": { "type": "object", "required": ["code", "name", "status"], "properties": { "code": { "$ref": "#/components/schemas/StableCode" }, "name": { "type": "string", "minLength": 1, "maxLength": 120 }, "status": { "type": "string", "enum": ["active", "suspended"] } }, "additionalProperties": false },
      "Product": { "type": "object", "required": ["product_id", "code", "name", "status", "provisioning_state", "official_tenant_id", "context_version", "created_at", "updated_at", "audit_id"], "properties": { "product_id": { "$ref": "#/components/schemas/Identifier" }, "code": { "$ref": "#/components/schemas/StableCode" }, "name": { "type": "string" }, "status": { "type": "string", "enum": ["active", "suspended"] }, "provisioning_state": { "type": "string", "const": "ready" }, "official_tenant_id": { "$ref": "#/components/schemas/Identifier" }, "context_version": { "type": "integer", "minimum": 1 }, "created_at": { "$ref": "#/components/schemas/Timestamp" }, "updated_at": { "$ref": "#/components/schemas/Timestamp" }, "audit_id": { "$ref": "#/components/schemas/Identifier" } }, "additionalProperties": false },
      "ProductSummary": { "type": "object", "required": ["product_id", "code", "name", "status", "provisioning_state", "context_version", "created_at", "updated_at"], "properties": { "product_id": { "$ref": "#/components/schemas/Identifier" }, "code": { "$ref": "#/components/schemas/StableCode" }, "name": { "type": "string" }, "status": { "type": "string", "enum": ["active", "suspended"] }, "provisioning_state": { "type": "string", "enum": ["pending", "ready", "failed"] }, "official_tenant_id": { "type": ["string", "null"], "maxLength": 128 }, "context_version": { "type": "integer", "minimum": 1 }, "created_at": { "$ref": "#/components/schemas/Timestamp" }, "updated_at": { "$ref": "#/components/schemas/Timestamp" } }, "additionalProperties": false },
...
      "MaskedUserIdentifier": { "type": "object", "required": ["type", "masked_value", "verified"], "properties": { "type": { "type": "string", "enum": ["email", "phone"] }, "masked_value": { "type": "string", "minLength": 1, "maxLength": 320 }, "verified": { "type": "boolean" } }, "additionalProperties": false },
      "ScopedUserAccess": { "type": "object", "required": ["scope_type", "scope_id", "status", "explicit", "version"], "properties": { "scope_type": { "type": "string", "enum": ["product", "tenant"] }, "scope_id": { "$ref": "#/components/schemas/Identifier" }, "status": { "type": "string", "enum": ["active", "suspended"] }, "explicit": { "type": "boolean" }, "version": { "type": "integer", "minimum": 0 }, "status_changed_at": { "type": ["string", "null"], "format": "date-time" } }, "additionalProperties": false },
      "AdminUserSummary": { "type": "object", "required": ["user_id", "user_version", "account_status", "display_name", "identifiers", "created_at", "member_since", "last_seen_at", "active_session_count", "total_session_count", "access"], "properties": { "user_id": { "$ref": "#/components/schemas/Identifier" }, "user_version": { "type": "integer", "minimum": 1 }, "account_status": { "type": "string", "enum": ["active", "locked", "disabled"] }, "display_name": { "type": ["string", "null"], "maxLength": 128 }, "identifiers": { "type": "array", "items": { "$ref": "#/components/schemas/MaskedUserIdentifier" } }, "created_at": { "$ref": "#/components/schemas/Timestamp" }, "member_since": { "type": ["string", "null"], "format": "date-time", "description": "Null only for the platform-global collection." }, "last_seen_at": { "type": ["string", "null"], "format": "date-time" }, "active_session_count": { "type": "integer", "minimum": 0 }, "total_session_count": { "type": "integer", "minimum": 0 }, "access": { "oneOf": [{ "$ref": "#/components/schemas/ScopedUserAccess" }, { "type": "null" }] } }, "additionalProperties": false },
...
      "UserPage": { "type": "object", "required": ["items", "next_cursor"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/AdminUserSummary" } }, "next_cursor": { "type": ["string", "null"] } }, "additionalProperties": false },
...
    }
  }
}